CAN TX writing to buffer
Added function to be able to write data of any data type to the data fields in a CAN TX buffer.
I think the initialization is correct but needs testing.
Added struct for a CAN message.
Changed variable type of buffers in global.h.
Need to write transmit function and test.